projects/ai-dashboard/backend/requirements.txt
# AI Dashboard Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Scheduling
apscheduler>=3.10.4

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# HTTP Client (let dependencies resolve version)
httpx>=0.25.1

# RSS Parsing
feedparser>=6.0.10

# AI/LLM
anthropic>=0.18.1
dspy-ai>=2.4.0

# Environment
python-dotenv>=1.0.0

# YouTube API
google-api-python-client>=2.118.0
google-auth-oauthlib>=1.2.0

# Apify
apify-client>=1.6.0

# Utilities
python-dateutil>=2.8.2
